Make gcc gcov first on path
lcov uses gcov, but there are two versions installed on the TravisCI osx images, the clang one and the gcc one. lcov needs the gcc one, but the clang one is (probably) first on the path, so by making the gcc one first on the path it should be chosen first.